remove utilsTest
