+++ src/lwc/geFormField/__tests__/geFormField.test.js
@@ -2,7 +2,6 @@ import { createElement } from 'lwc';
 import GeFormField from 'c/geFormField';
 import { registerLdsTestWireAdapter } from '@salesforce/sfdx-lwc-jest';
 import { getObjectInfo, getPicklistValues } from 'lightning/uiObjectInfoApi';
-import { shadowQuerySelector } from 'c/utilsTest';
 
 // Import mock data
 const mockGetPicklistValues = require('./data/wiredPicklistValues.json');
@@ -231,6 +230,39 @@ expect.extend({
         };
     }
 });
+
+/**
+ * Returns the shadowRoot property of a given Lightning web component.
+ * Originally sourced from lightning-global repo.
+ *
+ * @param {LWCElement} element The Lightning web component element to retrieve
+ * the shadowRoot property off of
+ * @returns {ShadowRoot} The shadow root of the given element
+ */
+export const getShadowRoot = (element) => {
+    if (!element || !element.shadowRoot) {
+        const tagName =
+            element && element.tagName && element.tagName.toLowerCase();
+        throw new Error(
+            `Attempting to retrieve the shadow root of '${tagName || element}'
+            but no shadowRoot property found`
+        );
+    }
+    return element.shadowRoot;
+}
+
+/**
+ * Non-recursively queries the template of the provided element using the
+ * provided selector.
+ *
+ * @param {LWCElement} element The Lightning web component element for which we 
+ * want to query the template.
+ * @param {String} selector The selector used to match the descendant element.
+ * @returns {Element}
+ */
+export const shadowQuerySelector = (element, selector) => {
+    return getShadowRoot(element).querySelector(selector);
+}
 
 const getSubComponentCombobox = geFormFieldElement => {
     return shadowQuerySelector(
